+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/KafkaConnectDockerfile.java
@@ -4,20 +4,26 @@
  */
 package io.strimzi.operator.cluster.model;
 
-import io.strimzi.api.kafka.model.connect.build.Artifact;
-import io.strimzi.api.kafka.model.connect.build.Build;
-import io.strimzi.api.kafka.model.connect.build.DownloadableArtifact;
-import io.strimzi.api.kafka.model.connect.build.JarArtifact;
-import io.strimzi.api.kafka.model.connect.build.OtherArtifact;
-import io.strimzi.api.kafka.model.connect.build.Plugin;
-import io.strimzi.api.kafka.model.connect.build.TgzArtifact;
-import io.strimzi.api.kafka.model.connect.build.ZipArtifact;
-import io.strimzi.operator.common.Util;
-
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
+
+import io.strimzi.api.kafka.model.connect.build.Artifact;
+import io.strimzi.api.kafka.model.connect.build.Build;
+import io.strimzi.api.kafka.model.connect.build.DownloadableArtifact;
+import io.strimzi.api.kafka.model.connect.build.JarArtifact;
+import io.strimzi.api.kafka.model.connect.build.MavenArtifact;
+import io.strimzi.api.kafka.model.connect.build.OtherArtifact;
+import io.strimzi.api.kafka.model.connect.build.Plugin;
+import io.strimzi.api.kafka.model.connect.build.TgzArtifact;
+import io.strimzi.api.kafka.model.connect.build.ZipArtifact;
+import io.strimzi.operator.cluster.ClusterOperatorConfig;
+import io.strimzi.operator.common.InvalidConfigurationException;
+import io.strimzi.operator.common.Util;
 
 /**
  * This class is used to generate the Dockerfile used by Kafka Connect Build. It takes the API definition with the
@@ -39,6 +45,74 @@ public class KafkaConnectDockerfile {
 
     private final String dockerfile;
 
+    private static final String DEFAULT_MAVEN_IMAGE = "quay.io/strimzi/maven-builder:latest";
+    private final String mavenBuilder;
+
+    public static Cmd run(String cmd, String... args) {
+        return new Cmd(new StringBuilder(), cmd, args);
+    }
+
+    public static class Cmd {
+
+        private final StringBuilder stringBuilder;
+        boolean doneFirst = false;
+
+        private Cmd(StringBuilder stringBuilder, String cmd, String... args) {
+            this.stringBuilder = stringBuilder;
+            append(cmd, args);
+        }
+
+        private Cmd append(String cmd, String[] args) {
+            append(cmd);
+            for (var s : args) {
+                append(s);
+            }
+            return this;
+        }
+
+        private Cmd append(String str) {
+            if (Objects.requireNonNull(str).isEmpty()) {
+                throw new IllegalArgumentException();
+            }
+            if (doneFirst) {
+                stringBuilder.append(' ');
+            }
+            doneFirst = true;
+
+            stringBuilder.append('\'');
+            for (var i = 0; i < str.length(); i++) {
+                char ch = str.charAt(i);
+                if (ch == '\'') {
+                    stringBuilder.append("'\"'\"'");
+                } else {
+                    stringBuilder.append(ch);
+                }
+            }
+            stringBuilder.append('\'');
+            return this;
+        }
+
+        public Cmd redirectTo(String str) {
+            stringBuilder.append(" >");
+            return append(str);
+        }
+
+        public Cmd pipeTo(String cmd, String... args) {
+            stringBuilder.append(" |");
+            return append(cmd, args);
+        }
+
+        public Cmd andRun(String cmd, String... args) {
+            stringBuilder.append(" \\\n      &&");
+            return append(cmd, args);
+        }
+
+        public String toString() {
+            return stringBuilder.toString();
+        }
+
+    }
+
     /**
      * Broker configuration template constructor
      *
@@ -46,10 +120,12 @@ public class KafkaConnectDockerfile {
      * @param connectBuild  The Build definition from the API
      */
     public KafkaConnectDockerfile(String fromImage, Build connectBuild) {
+        this.mavenBuilder = System.getenv().getOrDefault(ClusterOperatorConfig.STRIMZI_DEFAULT_MAVEN_BUILDER, DEFAULT_MAVEN_IMAGE);
         StringWriter stringWriter = new StringWriter();
         PrintWriter writer = new PrintWriter(stringWriter);
 
         printHeader(writer); // Print initial comment
+        connectorPluginsPreStage(writer, connectBuild.getPlugins());
         from(writer, fromImage); // Create FROM statement
         user(writer, ROOT_USER); // Switch to root user to be able to add plugins
         proxy(writer); // Configures proxy environment variables
@@ -59,6 +135,48 @@ public class KafkaConnectDockerfile {
         dockerfile = stringWriter.toString();
 
         writer.close();
+    }
+
+    /**
+     * Generates initial stage for multi-stage build
+     * @param writer        Writer for printing the Docker commands
+     * @param plugins       List of plugins which should be added to the container image
+     */
+    private void connectorPluginsPreStage(PrintWriter writer, List<Plugin> plugins) {
+        Map<String, List<MavenArtifact>> artifactMap = plugins.stream().collect(Collectors.toMap(plugin -> plugin.getName(),
+            plugin -> plugin.getArtifacts().stream().filter(artifact -> artifact instanceof MavenArtifact).map(artifact -> (MavenArtifact) artifact).collect(Collectors.toList())));
+        artifactMap.entrySet().removeIf(plugin -> plugin.getValue().isEmpty());
+
+        if (artifactMap.size() > 0) {
+            writer.println("FROM " + mavenBuilder + " AS downloadArtifacts");
+            artifactMap.entrySet().forEach(plugin -> {
+                plugin.getValue().forEach(mvn -> {
+                    String repo = mvn.getRepository() == null ? MavenArtifact.DEFAULT_REPOSITORY : maybeAppendSlash(mvn.getRepository());
+                    String artifactHash = Util.sha1Prefix(mvn.getGroup() + "/" + mvn.getArtifact() + "/" + mvn.getVersion());
+                    String artifactDir = plugin.getKey() + "/" + artifactHash;
+
+                    Cmd cmd = run("curl", "-L", "--create-dirs", "--output", "/tmp/" + artifactDir + "/pom.xml", assembleResourceUrl(repo, mvn, "pom"))
+                            .andRun("mvn", "dependency:copy-dependencies",
+                                    "-DoutputDirectory=/tmp/artifacts/" + artifactDir, "-f", "/tmp/" + artifactDir + "/pom.xml")
+                            .andRun("curl", "-L", "--create-dirs", "--output",
+                                    "/tmp/artifacts/" + artifactDir + "/" + mvn.getArtifact() + "-" + mvn.getVersion() + ".jar",
+                                    assembleResourceUrl(repo, mvn, "jar"));
+                    writer.append("RUN ").println(cmd);
+                    writer.println();
+                });
+            });
+        }
+    }
+
+    private String assembleResourceUrl(String repo, MavenArtifact mvn, String extension) {
+        return String.format("%s%s/%s/%s/%s-%s.%s",
+                repo,
+                mvn.getGroup().replace(".", "/"),
+                mvn.getArtifact().replace(".", "/"),
+                mvn.getVersion(),
+                mvn.getArtifact(),
+                mvn.getVersion(),
+                extension);
     }
 
     /**
@@ -106,7 +224,7 @@ public class KafkaConnectDockerfile {
     }
 
     /**
-     * Adds the commands to donwload and possibly unpact the connector plugins
+     * Adds the commands to download and possibly unpack the connector plugins
      *
      * @param writer    Writer for printing the Docker commands
      * @param plugins   List of plugins which should be added to the container image
@@ -136,11 +254,31 @@ public class KafkaConnectDockerfile {
                 addTgzArtifact(writer, connectorPath, (TgzArtifact) art);
             } else if (art instanceof ZipArtifact) {
                 addZipArtifact(writer, connectorPath, (ZipArtifact) art);
+            } else if (art instanceof MavenArtifact) {
+                addMavenArtifact(writer, plugin.getName(), (MavenArtifact) art);
             } else if (art instanceof OtherArtifact) {
                 addOtherArtifact(writer, connectorPath, (OtherArtifact) art);
             } else {
                 throw new RuntimeException("Unexpected artifact type " + art.getType());
             }
+        }
+    }
+
+    private void checkUrlIsPresent(DownloadableArtifact art) {
+        if (art.getUrl() == null || art.getUrl().isEmpty()) {
+            throw new InvalidConfigurationException("`" + art.getType() + "` artifact is missing a URL.");
+        }
+    }
+
+    private void checkGavIsPresent(MavenArtifact art) {
+        if (art.getGroup() == null || art.getGroup().isEmpty()) {
+            throw new InvalidConfigurationException("`" + art.getType() + "` artifact is missing a group ID.");
+        }
+        if (art.getArtifact() == null || art.getArtifact().isEmpty()) {
+            throw new InvalidConfigurationException("`" + art.getType() + "` artifact is missing an artifact ID.");
+        }
+        if (art.getVersion() == null || art.getVersion().isEmpty()) {
+            throw new InvalidConfigurationException("`" + art.getType() + "` artifact is missing a version number.");
         }
     }
 
@@ -152,12 +290,11 @@ public class KafkaConnectDockerfile {
      * @param jar               The JAR-type artifact
      */
     private void addJarArtifact(PrintWriter writer, String connectorPath, JarArtifact jar) {
+        checkUrlIsPresent(jar);
         String artifactHash = Util.sha1Prefix(jar.getUrl());
         String artifactDir = connectorPath + "/" + artifactHash;
         String artifactPath = artifactDir + "/" + artifactHash + ".jar";
-        String downloadCmd =  "curl -L --output " + artifactPath + " " + jar.getUrl();
-
-        addUnmodifiedArtifact(writer, jar, artifactDir, downloadCmd, artifactPath);
+        addUnmodifiedArtifact(writer, jar, artifactDir, artifactPath);
     }
 
     /**
@@ -168,40 +305,35 @@ public class KafkaConnectDockerfile {
      * @param other             The Other-type artifact
      */
     private void addOtherArtifact(PrintWriter writer, String connectorPath, OtherArtifact other) {
+        checkUrlIsPresent(other);
         String artifactHash = Util.sha1Prefix(other.getUrl());
         String artifactDir = connectorPath + "/" + artifactHash;
         String fileName = other.getFileName() != null ? other.getFileName() : artifactHash;
         String artifactPath = artifactDir + "/" + fileName;
-        String downloadCmd =  "curl -L --output " + artifactPath + " " + other.getUrl();
 
-        addUnmodifiedArtifact(writer, other, artifactDir, downloadCmd, artifactPath);
+        addUnmodifiedArtifact(writer, other, artifactDir, artifactPath);
     }
 
     /**
      * Adds download command for artifacts which are just downloaded without any unpacking or other processing.
-     *
      * @param writer            Writer for printing the Docker commands
      * @param art               Artifact which should be downloaded
      * @param artifactDir       Directory into which the artifact should be downloaded
-     * @param downloadCmd       Command for downloading the artifact
      * @param artifactPath      Full path of the artifact
      */
-    private void addUnmodifiedArtifact(PrintWriter writer, DownloadableArtifact art, String artifactDir, String downloadCmd, String artifactPath)    {
-        writer.println("RUN mkdir -p " + artifactDir + " \\");
+    private void addUnmodifiedArtifact(PrintWriter writer, DownloadableArtifact art, String artifactDir, String artifactPath) {
+        Cmd run = run("mkdir", "-p", artifactDir)
+                .andRun("curl", "-L", "--output", artifactPath, art.getUrl());
 
-        if (art.getSha512sum() == null || art.getSha512sum().isEmpty()) {
-            // No checksum => we just download the file
-            writer.println("      && " + downloadCmd);
-        } else {
+        if (art.getSha512sum() != null && !art.getSha512sum().isEmpty()) {
             // Checksum exists => we need to check it
-            String checksum = art.getSha512sum() + " " + artifactPath;
-
-            writer.println("      && " + downloadCmd + " \\");
-            writer.println("      && echo \"" + checksum + "\" > " + artifactPath + ".sha512 \\");
-            writer.println("      && sha512sum --check " + artifactPath + ".sha512 \\");
-            writer.println("      && rm -f " + artifactPath + ".sha512");
+            String shaFile = artifactPath + ".sha512";
+            run.andRun("echo", art.getSha512sum() + " " + artifactPath)
+                    .redirectTo(shaFile)
+                .andRun("sha512sum", "--check", shaFile)
+                .andRun("rm", "-f", shaFile);
         }
-
+        writer.append("RUN ").println(run);
         writer.println();
     }
 
@@ -213,33 +345,25 @@ public class KafkaConnectDockerfile {
      * @param tgz               The TGZ-type artifact
      */
     private void addTgzArtifact(PrintWriter writer, String connectorPath, TgzArtifact tgz) {
+        checkUrlIsPresent(tgz);
         String artifactHash = Util.sha1Prefix(tgz.getUrl());
         String artifactDir = connectorPath + "/" + artifactHash;
         String archivePath = connectorPath + "/" + artifactHash + ".tgz";
 
-        String downloadCmd =  "curl -L --output " + archivePath + " " + tgz.getUrl();
-        String unpackCmd =  "tar xvfz " + archivePath + " -C " + artifactDir;
-        String deleteCmd =  "rm -vf " + archivePath;
+        Cmd run = run("mkdir", "-p", artifactDir)
+                .andRun("curl", "-L", "--output", archivePath, tgz.getUrl());
 
-        writer.println("RUN mkdir -p " + artifactDir + " \\");
-
-        if (tgz.getSha512sum() == null || tgz.getSha512sum().isEmpty()) {
-            // No checksum => we just download and unpack the file
-            writer.println("      && " + downloadCmd + " \\");
-            writer.println("      && " + unpackCmd + " \\");
-            writer.println("      && " + deleteCmd);
-        } else {
+        if (tgz.getSha512sum() != null && !tgz.getSha512sum().isEmpty()) {
             // Checksum exists => we need to check it
-            String checksum = tgz.getSha512sum() + " " + archivePath;
-
-            writer.println("      && " + downloadCmd + " \\");
-            writer.println("      && echo \"" + checksum + "\" > " + archivePath + ".sha512 \\");
-            writer.println("      && sha512sum --check " + archivePath + ".sha512 \\");
-            writer.println("      && rm -f " + archivePath + ".sha512 \\");
-            writer.println("      && " + unpackCmd + " \\");
-            writer.println("      && " + deleteCmd);
+            String shaFile = archivePath + ".sha512";
+            run.andRun("echo", tgz.getSha512sum() + " " + archivePath)
+                    .redirectTo(shaFile)
+                .andRun("sha512sum", "--check", shaFile)
+                .andRun("rm", "-f", shaFile);
         }
-
+        run.andRun("tar", "xvfz", archivePath, "-C", artifactDir)
+            .andRun("rm", "-vf", archivePath);
+        writer.append("RUN ").println(run);
         writer.println();
     }
 
@@ -251,37 +375,57 @@ public class KafkaConnectDockerfile {
      * @param zip               The ZIP-type artifact
      */
     private void addZipArtifact(PrintWriter writer, String connectorPath, ZipArtifact zip) {
+        checkUrlIsPresent(zip);
         String artifactHash = Util.sha1Prefix(zip.getUrl());
         String artifactDir = connectorPath + "/" + artifactHash;
         String archivePath = connectorPath + "/" + artifactHash + ".zip";
 
-        String downloadCmd =  "curl -L --output " + archivePath + " " + zip.getUrl();
-        String unpackCmd =  "unzip " + archivePath + " -d " + artifactDir;
-        String deleteSymLinks = "find " + artifactDir + " -type l | xargs rm -f";
-        String deleteCmd =  "rm -vf " + archivePath;
+        Cmd run = run("mkdir", "-p", artifactDir)
+                .andRun("curl", "-L", "--output", archivePath, zip.getUrl());
 
-        writer.println("RUN mkdir -p " + artifactDir + " \\");
-
-        if (zip.getSha512sum() == null || zip.getSha512sum().isEmpty()) {
-            // No checksum => we just download and unpack the file
-            writer.println("      && " + downloadCmd + " \\");
-            writer.println("      && " + unpackCmd + " \\");
-            writer.println("      && " + deleteSymLinks + " \\");
-            writer.println("      && " + deleteCmd);
-        } else {
+        if (zip.getSha512sum() != null && !zip.getSha512sum().isEmpty()) {
             // Checksum exists => we need to check it
-            String checksum = zip.getSha512sum() + " " + archivePath;
-
-            writer.println("      && " + downloadCmd + " \\");
-            writer.println("      && echo \"" + checksum + "\" > " + archivePath + ".sha512 \\");
-            writer.println("      && sha512sum --check " + archivePath + ".sha512 \\");
-            writer.println("      && rm -f " + archivePath + ".sha512 \\");
-            writer.println("      && " + unpackCmd + " \\");
-            writer.println("      && " + deleteSymLinks + " \\");
-            writer.println("      && " + deleteCmd);
+            String shaFile = archivePath + ".sha512";
+            run.andRun("echo", zip.getSha512sum() + " " + archivePath)
+                    .redirectTo(shaFile)
+                .andRun("sha512sum", "--check", shaFile)
+                .andRun("rm", "-f", shaFile);
         }
 
+        run.andRun("unzip", archivePath, "-d", artifactDir)
+            .andRun("find", artifactDir, "-type", "l").pipeTo("xargs", "rm", "-f")
+            .andRun("rm", "-vf", archivePath);
+
+        writer.append("RUN ").println(run);
         writer.println();
+    }
+
+    /**
+     * Add command sequence for downloading Maven artifact
+     *
+     * @param writer            Writer for printing the Docker commands
+     * @param connectorName     Name of the connector to which this artifact belongs should be downloaded
+     * @param mvn               The maven artifact
+     */
+    private void addMavenArtifact(PrintWriter writer, String connectorName, MavenArtifact mvn) {
+        checkGavIsPresent(mvn);
+        String artifactHash = Util.sha1Prefix(mvn.getGroup() + "/" + mvn.getArtifact() + "/" + mvn.getVersion());
+
+        Cmd run = run("/tmp/artifacts/" + connectorName + "/" + artifactHash, BASE_PLUGIN_PATH + connectorName + "/" + artifactHash);
+        writer.append("COPY --from=downloadArtifacts ").println(run);
+        writer.println();
+    }
+
+    /**
+     * @param checked The string to check whether contains the slash as the last character
+     * @return The string with slash ('/') as the last character
+     */
+    private String maybeAppendSlash(String checked) {
+        if (checked.lastIndexOf('/') + 1 == checked.length()) {
+            return checked;
+        } else {
+            return checked + "/";
+        }
     }
 
     /**
@@ -325,4 +469,6 @@ public class KafkaConnectDockerfile {
     public String hashStub()    {
         return Util.sha1Prefix(dockerfile);
     }
+
+
 }