+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/KafkaConnectDockerfileTest.java
@@ -7,18 +7,24 @@ package io.strimzi.operator.cluster.model;
 import io.strimzi.api.kafka.model.connect.build.Artifact;
 import io.strimzi.api.kafka.model.connect.build.Build;
 import io.strimzi.api.kafka.model.connect.build.BuildBuilder;
+import io.strimzi.api.kafka.model.connect.build.JarArtifact;
 import io.strimzi.api.kafka.model.connect.build.JarArtifactBuilder;
+import io.strimzi.api.kafka.model.connect.build.MavenArtifact;
+import io.strimzi.api.kafka.model.connect.build.MavenArtifactBuilder;
 import io.strimzi.api.kafka.model.connect.build.OtherArtifactBuilder;
 import io.strimzi.api.kafka.model.connect.build.PluginBuilder;
 import io.strimzi.api.kafka.model.connect.build.TgzArtifactBuilder;
 import io.strimzi.api.kafka.model.connect.build.ZipArtifactBuilder;
+import io.strimzi.operator.common.InvalidConfigurationException;
 import io.strimzi.test.annotations.ParallelSuite;
 import io.strimzi.test.annotations.ParallelTest;
 
 import static io.strimzi.operator.cluster.model.KafkaBrokerConfigurationBuilderTest.IsEquivalent.isEquivalent;
 import static java.util.Collections.emptyList;
+import static java.util.Collections.singletonList;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @ParallelSuite
 public class KafkaConnectDockerfileTest {
@@ -107,8 +113,8 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/51e5038c \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/51e5038c/51e5038c.jar https://mydomain.tld/my.jar",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/51e5038c' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/51e5038c/51e5038c.jar' 'https://mydomain.tld/my.jar'",
                 "USER 1001"));
     }
 
@@ -125,11 +131,11 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/0df6d15c \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar https://mydomain.tld/my2.jar \\",
-                "      && echo \"sha-512-checksum /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar\" > /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512 \\",
-                "      && sha512sum --check /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512 \\",
-                "      && rm -f /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/0df6d15c' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar' 'https://mydomain.tld/my2.jar' \\",
+                "      && 'echo' 'sha-512-checksum /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar' > '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512' \\",
+                "      && 'sha512sum' '--check' '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512' \\",
+                "      && 'rm' '-f' '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512'",
                 "USER 1001"));
     }
 
@@ -146,8 +152,8 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/2c3b64c7 \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/2c3b64c7/my.so https://mydomain.tld/download?artifact=my.so",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/2c3b64c7' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/2c3b64c7/my.so' 'https://mydomain.tld/download?artifact=my.so'",
                 "USER 1001"));
     }
 
@@ -164,11 +170,11 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/2e6fee06 \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/2e6fee06/my2.so https://mydomain.tld/download?artifactId=1874 \\",
-                "      && echo \"sha-512-checksum /opt/kafka/plugins/my-connector-plugin/2e6fee06/my2.so\" > /opt/kafka/plugins/my-connector-plugin/2e6fee06/my2.so.sha512 \\",
-                "      && sha512sum --check /opt/kafka/plugins/my-connector-plugin/2e6fee06/my2.so.sha512 \\",
-                "      && rm -f /opt/kafka/plugins/my-connector-plugin/2e6fee06/my2.so.sha512",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/2e6fee06' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/2e6fee06/my2.so' 'https://mydomain.tld/download?artifactId=1874' \\",
+                "      && 'echo' 'sha-512-checksum /opt/kafka/plugins/my-connector-plugin/2e6fee06/my2.so' > '/opt/kafka/plugins/my-connector-plugin/2e6fee06/my2.so.sha512' \\",
+                "      && 'sha512sum' '--check' '/opt/kafka/plugins/my-connector-plugin/2e6fee06/my2.so.sha512' \\",
+                "      && 'rm' '-f' '/opt/kafka/plugins/my-connector-plugin/2e6fee06/my2.so.sha512'",
                 "USER 1001"));
     }
 
@@ -185,11 +191,11 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/2e6fee06 \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/2e6fee06/2e6fee06 https://mydomain.tld/download?artifactId=1874 \\",
-                "      && echo \"sha-512-checksum /opt/kafka/plugins/my-connector-plugin/2e6fee06/2e6fee06\" > /opt/kafka/plugins/my-connector-plugin/2e6fee06/2e6fee06.sha512 \\",
-                "      && sha512sum --check /opt/kafka/plugins/my-connector-plugin/2e6fee06/2e6fee06.sha512 \\",
-                "      && rm -f /opt/kafka/plugins/my-connector-plugin/2e6fee06/2e6fee06.sha512",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/2e6fee06' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/2e6fee06/2e6fee06' 'https://mydomain.tld/download?artifactId=1874' \\",
+                "      && 'echo' 'sha-512-checksum /opt/kafka/plugins/my-connector-plugin/2e6fee06/2e6fee06' > '/opt/kafka/plugins/my-connector-plugin/2e6fee06/2e6fee06.sha512' \\",
+                "      && 'sha512sum' '--check' '/opt/kafka/plugins/my-connector-plugin/2e6fee06/2e6fee06.sha512' \\",
+                "      && 'rm' '-f' '/opt/kafka/plugins/my-connector-plugin/2e6fee06/2e6fee06.sha512'",
                 "USER 1001"));
     }
 
@@ -206,13 +212,13 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/51e5038c \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/51e5038c/51e5038c.jar https://mydomain.tld/my.jar",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/0df6d15c \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar https://mydomain.tld/my2.jar \\",
-                "      && echo \"sha-512-checksum /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar\" > /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512 \\",
-                "      && sha512sum --check /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512 \\",
-                "      && rm -f /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/51e5038c' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/51e5038c/51e5038c.jar' 'https://mydomain.tld/my.jar'",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/0df6d15c' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar' 'https://mydomain.tld/my2.jar' \\",
+                "      && 'echo' 'sha-512-checksum /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar' > '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512' \\",
+                "      && 'sha512sum' '--check' '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512' \\",
+                "      && 'rm' '-f' '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512'",
                 "USER 1001"));
     }
 
@@ -229,10 +235,10 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/6718766b \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/6718766b.tgz https://mydomain.tld/my.tgz \\",
-                "      && tar xvfz /opt/kafka/plugins/my-connector-plugin/6718766b.tgz -C /opt/kafka/plugins/my-connector-plugin/6718766b \\",
-                "      && rm -vf /opt/kafka/plugins/my-connector-plugin/6718766b.tgz",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/6718766b' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/6718766b.tgz' 'https://mydomain.tld/my.tgz' \\",
+                "      && 'tar' 'xvfz' '/opt/kafka/plugins/my-connector-plugin/6718766b.tgz' '-C' '/opt/kafka/plugins/my-connector-plugin/6718766b' \\",
+                "      && 'rm' '-vf' '/opt/kafka/plugins/my-connector-plugin/6718766b.tgz'",
                 "USER 1001"));
     }
 
@@ -249,11 +255,11 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/d8d533bc \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/d8d533bc.zip https://mydomain.tld/my.zip \\",
-                "      && unzip /opt/kafka/plugins/my-connector-plugin/d8d533bc.zip -d /opt/kafka/plugins/my-connector-plugin/d8d533bc \\",
-                "      && find /opt/kafka/plugins/my-connector-plugin/d8d533bc -type l | xargs rm -f \\",
-                "      && rm -vf /opt/kafka/plugins/my-connector-plugin/d8d533bc.zip",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/d8d533bc' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/d8d533bc.zip' 'https://mydomain.tld/my.zip' \\",
+                "      && 'unzip' '/opt/kafka/plugins/my-connector-plugin/d8d533bc.zip' '-d' '/opt/kafka/plugins/my-connector-plugin/d8d533bc' \\",
+                "      && 'find' '/opt/kafka/plugins/my-connector-plugin/d8d533bc' '-type' 'l' | 'xargs' 'rm' '-f' \\",
+                "      && 'rm' '-vf' '/opt/kafka/plugins/my-connector-plugin/d8d533bc.zip'",
                 "USER 1001"));
     }
 
@@ -270,14 +276,14 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/90e04094 \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/90e04094.zip https://mydomain.tld/my2.zip \\",
-                "      && echo \"sha-512-checksum /opt/kafka/plugins/my-connector-plugin/90e04094.zip\" > /opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512 \\",
-                "      && sha512sum --check /opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512 \\",
-                "      && rm -f /opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512 \\",
-                "      && unzip /opt/kafka/plugins/my-connector-plugin/90e04094.zip -d /opt/kafka/plugins/my-connector-plugin/90e04094 \\",
-                "      && find /opt/kafka/plugins/my-connector-plugin/90e04094 -type l | xargs rm -f \\",
-                "      && rm -vf /opt/kafka/plugins/my-connector-plugin/90e04094.zip",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/90e04094' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/90e04094.zip' 'https://mydomain.tld/my2.zip' \\",
+                "      && 'echo' 'sha-512-checksum /opt/kafka/plugins/my-connector-plugin/90e04094.zip' > '/opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512' \\",
+                "      && 'sha512sum' '--check' '/opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512' \\",
+                "      && 'rm' '-f' '/opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512' \\",
+                "      && 'unzip' '/opt/kafka/plugins/my-connector-plugin/90e04094.zip' '-d' '/opt/kafka/plugins/my-connector-plugin/90e04094' \\",
+                "      && 'find' '/opt/kafka/plugins/my-connector-plugin/90e04094' '-type' 'l' | 'xargs' 'rm' '-f' \\",
+                "      && 'rm' '-vf' '/opt/kafka/plugins/my-connector-plugin/90e04094.zip'",
                 "USER 1001"));
     }
 
@@ -294,13 +300,13 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/638bd501 \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/638bd501.tgz https://mydomain.tld/my2.tgz \\",
-                "      && echo \"sha-512-checksum /opt/kafka/plugins/my-connector-plugin/638bd501.tgz\" > /opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512 \\",
-                "      && sha512sum --check /opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512 \\",
-                "      && rm -f /opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512 \\",
-                "      && tar xvfz /opt/kafka/plugins/my-connector-plugin/638bd501.tgz -C /opt/kafka/plugins/my-connector-plugin/638bd501 \\",
-                "      && rm -vf /opt/kafka/plugins/my-connector-plugin/638bd501.tgz",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/638bd501' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz' 'https://mydomain.tld/my2.tgz' \\",
+                "      && 'echo' 'sha-512-checksum /opt/kafka/plugins/my-connector-plugin/638bd501.tgz' > '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512' \\",
+                "      && 'sha512sum' '--check' '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512' \\",
+                "      && 'rm' '-f' '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512' \\",
+                "      && 'tar' 'xvfz' '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz' '-C' '/opt/kafka/plugins/my-connector-plugin/638bd501' \\",
+                "      && 'rm' '-vf' '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz'",
                 "USER 1001"));
     }
 
@@ -317,17 +323,17 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/6718766b \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/6718766b.tgz https://mydomain.tld/my.tgz \\",
-                "      && tar xvfz /opt/kafka/plugins/my-connector-plugin/6718766b.tgz -C /opt/kafka/plugins/my-connector-plugin/6718766b \\",
-                "      && rm -vf /opt/kafka/plugins/my-connector-plugin/6718766b.tgz",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/638bd501 \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/638bd501.tgz https://mydomain.tld/my2.tgz \\",
-                "      && echo \"sha-512-checksum /opt/kafka/plugins/my-connector-plugin/638bd501.tgz\" > /opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512 \\",
-                "      && sha512sum --check /opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512 \\",
-                "      && rm -f /opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512 \\",
-                "      && tar xvfz /opt/kafka/plugins/my-connector-plugin/638bd501.tgz -C /opt/kafka/plugins/my-connector-plugin/638bd501 \\",
-                "      && rm -vf /opt/kafka/plugins/my-connector-plugin/638bd501.tgz",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/6718766b' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/6718766b.tgz' 'https://mydomain.tld/my.tgz' \\",
+                "      && 'tar' 'xvfz' '/opt/kafka/plugins/my-connector-plugin/6718766b.tgz' '-C' '/opt/kafka/plugins/my-connector-plugin/6718766b' \\",
+                "      && 'rm' '-vf' '/opt/kafka/plugins/my-connector-plugin/6718766b.tgz'",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/638bd501' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz' 'https://mydomain.tld/my2.tgz' \\",
+                "      && 'echo' 'sha-512-checksum /opt/kafka/plugins/my-connector-plugin/638bd501.tgz' > '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512' \\",
+                "      && 'sha512sum' '--check' '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512' \\",
+                "      && 'rm' '-f' '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz.sha512' \\",
+                "      && 'tar' 'xvfz' '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz' '-C' '/opt/kafka/plugins/my-connector-plugin/638bd501' \\",
+                "      && 'rm' '-vf' '/opt/kafka/plugins/my-connector-plugin/638bd501.tgz'",
                 "USER 1001"));
     }
 
@@ -344,19 +350,19 @@ public class KafkaConnectDockerfileTest {
 
         assertThat(df.getDockerfile(), isEquivalent("FROM myImage:latest",
                 "USER root:root",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/d8d533bc \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/d8d533bc.zip https://mydomain.tld/my.zip \\",
-                "      && unzip /opt/kafka/plugins/my-connector-plugin/d8d533bc.zip -d /opt/kafka/plugins/my-connector-plugin/d8d533bc \\",
-                "      && find /opt/kafka/plugins/my-connector-plugin/d8d533bc -type l | xargs rm -f \\",
-                "      && rm -vf /opt/kafka/plugins/my-connector-plugin/d8d533bc.zip",
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/90e04094 \\",
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/90e04094.zip https://mydomain.tld/my2.zip \\",
-                "      && echo \"sha-512-checksum /opt/kafka/plugins/my-connector-plugin/90e04094.zip\" > /opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512 \\",
-                "      && sha512sum --check /opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512 \\",
-                "      && rm -f /opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512 \\",
-                "      && unzip /opt/kafka/plugins/my-connector-plugin/90e04094.zip -d /opt/kafka/plugins/my-connector-plugin/90e04094 \\",
-                "      && find /opt/kafka/plugins/my-connector-plugin/90e04094 -type l | xargs rm -f \\",
-                "      && rm -vf /opt/kafka/plugins/my-connector-plugin/90e04094.zip",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/d8d533bc' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/d8d533bc.zip' 'https://mydomain.tld/my.zip' \\",
+                "      && 'unzip' '/opt/kafka/plugins/my-connector-plugin/d8d533bc.zip' '-d' '/opt/kafka/plugins/my-connector-plugin/d8d533bc' \\",
+                "      && 'find' '/opt/kafka/plugins/my-connector-plugin/d8d533bc' '-type' 'l' | 'xargs' 'rm' '-f' \\",
+                "      && 'rm' '-vf' '/opt/kafka/plugins/my-connector-plugin/d8d533bc.zip'",
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/90e04094' \\",
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/90e04094.zip' 'https://mydomain.tld/my2.zip' \\",
+                "      && 'echo' 'sha-512-checksum /opt/kafka/plugins/my-connector-plugin/90e04094.zip' > '/opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512' \\",
+                "      && 'sha512sum' '--check' '/opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512' \\",
+                "      && 'rm' '-f' '/opt/kafka/plugins/my-connector-plugin/90e04094.zip.sha512' \\",
+                "      && 'unzip' '/opt/kafka/plugins/my-connector-plugin/90e04094.zip' '-d' '/opt/kafka/plugins/my-connector-plugin/90e04094' \\",
+                "      && 'find' '/opt/kafka/plugins/my-connector-plugin/90e04094' '-type' 'l' | 'xargs' 'rm' '-f' \\",
+                "      && 'rm' '-vf' '/opt/kafka/plugins/my-connector-plugin/90e04094.zip'",
                 "USER 1001"));
     }
 
@@ -385,15 +391,157 @@ public class KafkaConnectDockerfileTest {
                 "##########\n" +
                 "# Connector plugin my-connector-plugin\n" +
                 "##########\n" +
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/51e5038c \\\n" +
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/51e5038c/51e5038c.jar https://mydomain.tld/my.jar\n" +
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/51e5038c' \\\n" +
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/51e5038c/51e5038c.jar' 'https://mydomain.tld/my.jar'\n" +
                 "\n" +
-                "RUN mkdir -p /opt/kafka/plugins/my-connector-plugin/0df6d15c \\\n" +
-                "      && curl -L --output /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar https://mydomain.tld/my2.jar \\\n" +
-                "      && echo \"sha-512-checksum /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar\" > /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512 \\\n" +
-                "      && sha512sum --check /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512 \\\n" +
-                "      && rm -f /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512\n" +
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/0df6d15c' \\\n" +
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar' 'https://mydomain.tld/my2.jar' \\\n" +
+                "      && 'echo' 'sha-512-checksum /opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar' > '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512' \\\n" +
+                "      && 'sha512sum' '--check' '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512' \\\n" +
+                "      && 'rm' '-f' '/opt/kafka/plugins/my-connector-plugin/0df6d15c/0df6d15c.jar.sha512'\n" +
                 "\n" +
                 "USER 1001\n\n"));
+    }
+
+    @ParallelTest
+    public void testNoUrlWhenRequired() {
+        Artifact tgzArtifact = new TgzArtifactBuilder()
+                .build();
+        Artifact jarArtifact = new JarArtifactBuilder()
+                .build();
+        Artifact zipArtifact = new ZipArtifactBuilder()
+                .build();
+        Artifact otherArtifact = new OtherArtifactBuilder()
+                .build();
+
+        Build connectBuildTgz = new BuildBuilder()
+                .withPlugins(new PluginBuilder()
+                        .withName("my-connector-plugin")
+                        .withArtifacts(singletonList(tgzArtifact))
+                        .build())
+                .build();
+
+        Throwable e = assertThrows(InvalidConfigurationException.class, () -> new KafkaConnectDockerfile("myImage:latest", connectBuildTgz));
+        assertThat(e.getMessage(), is("`tgz` artifact is missing a URL."));
+
+        Build connectBuildZip = new BuildBuilder()
+                .withPlugins(new PluginBuilder()
+                        .withName("my-connector-plugin")
+                        .withArtifacts(singletonList(zipArtifact))
+                        .build())
+                .build();
+
+        e = assertThrows(InvalidConfigurationException.class, () -> new KafkaConnectDockerfile("myImage:latest", connectBuildZip));
+        assertThat(e.getMessage(), is("`zip` artifact is missing a URL."));
+
+        Build connectBuildJar = new BuildBuilder()
+                .withPlugins(new PluginBuilder()
+                        .withName("my-connector-plugin")
+                        .withArtifacts(singletonList(jarArtifact))
+                        .build())
+                .build();
+
+        e = assertThrows(InvalidConfigurationException.class, () -> new KafkaConnectDockerfile("myImage:latest", connectBuildJar));
+        assertThat(e.getMessage(), is("`jar` artifact is missing a URL."));
+
+        Build connectBuildOther = new BuildBuilder()
+                .withPlugins(new PluginBuilder()
+                        .withName("my-connector-plugin")
+                        .withArtifacts(singletonList(otherArtifact))
+                        .build())
+                .build();
+
+        e = assertThrows(InvalidConfigurationException.class, () -> new KafkaConnectDockerfile("myImage:latest", connectBuildOther));
+        assertThat(e.getMessage(), is("`other` artifact is missing a URL."));
+    }
+
+    @ParallelTest
+    public void testMavenDockerfile()   {
+        JarArtifact jar = new JarArtifactBuilder()
+                .withUrl("http://url.com/ar.jar")
+                .build();
+
+        MavenArtifact mvn1 = new MavenArtifactBuilder()
+                .withGroup("g1")
+                .withArtifact("a1")
+                .withVersion("v1")
+                .build();
+
+        MavenArtifact mvn2 = new MavenArtifactBuilder()
+                .withGroup("g2")
+                .withArtifact("a2")
+                .withVersion("v2")
+                .build();
+
+        Build connectBuild = new BuildBuilder()
+                .withPlugins(
+                    new PluginBuilder()
+                        .withName("my-connector-plugin")
+                        .withArtifacts(jar, mvn1, mvn2)
+                        .build(),
+                    new PluginBuilder()
+                        .withName("other-connector-plugin")
+                        .withArtifacts(jar)
+                        .build())
+                .build();
+
+        KafkaConnectDockerfile df = new KafkaConnectDockerfile("myImage:latest", connectBuild);
+
+        assertThat(df.getDockerfile(), is("##############################\n" +
+                "##############################\n" +
+                "# This file is automatically generated by the Strimzi Cluster Operator\n" +
+                "# Any changes to this file will be ignored and overwritten!\n" +
+                "##############################\n" +
+                "##############################\n" +
+                "\n" +
+                "FROM quay.io/strimzi/maven-builder:latest AS downloadArtifacts\n" +
+                "RUN 'curl' '-L' '--create-dirs' '--output' '/tmp/my-connector-plugin/64cebd9c/pom.xml' 'https://repo1.maven.org/maven2/g1/a1/v1/a1-v1.pom' \\\n" +
+                "      && 'mvn' 'dependency:copy-dependencies' '-DoutputDirectory=/tmp/artifacts/my-connector-plugin/64cebd9c' '-f' '/tmp/my-connector-plugin/64cebd9c/pom.xml' \\\n" +
+                "      && 'curl' '-L' '--create-dirs' '--output' '/tmp/artifacts/my-connector-plugin/64cebd9c/a1-v1.jar' 'https://repo1.maven.org/maven2/g1/a1/v1/a1-v1.jar'\n" +
+                "\n" +
+                "RUN 'curl' '-L' '--create-dirs' '--output' '/tmp/my-connector-plugin/9983060e/pom.xml' 'https://repo1.maven.org/maven2/g2/a2/v2/a2-v2.pom' \\\n" +
+                "      && 'mvn' 'dependency:copy-dependencies' '-DoutputDirectory=/tmp/artifacts/my-connector-plugin/9983060e' '-f' '/tmp/my-connector-plugin/9983060e/pom.xml' \\\n" +
+                "      && 'curl' '-L' '--create-dirs' '--output' '/tmp/artifacts/my-connector-plugin/9983060e/a2-v2.jar' 'https://repo1.maven.org/maven2/g2/a2/v2/a2-v2.jar'\n" +
+                "\n" +
+                "FROM myImage:latest\n" +
+                "\n" +
+                "USER root:root\n" +
+                "\n" +
+                "##########\n" +
+                "# Connector plugin my-connector-plugin\n" +
+                "##########\n" +
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/my-connector-plugin/9bb2fd11' \\\n" +
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/my-connector-plugin/9bb2fd11/9bb2fd11.jar' 'http://url.com/ar.jar'\n" +
+                "\n" +
+                "COPY --from=downloadArtifacts '/tmp/artifacts/my-connector-plugin/64cebd9c' '/opt/kafka/plugins/my-connector-plugin/64cebd9c'\n" +
+                "\n" +
+                "COPY --from=downloadArtifacts '/tmp/artifacts/my-connector-plugin/9983060e' '/opt/kafka/plugins/my-connector-plugin/9983060e'\n" +
+                "\n" +
+                "##########\n" +
+                "# Connector plugin other-connector-plugin\n" +
+                "##########\n" +
+                "RUN 'mkdir' '-p' '/opt/kafka/plugins/other-connector-plugin/9bb2fd11' \\\n" +
+                "      && 'curl' '-L' '--output' '/opt/kafka/plugins/other-connector-plugin/9bb2fd11/9bb2fd11.jar' 'http://url.com/ar.jar'\n" +
+                "\n" +
+                "USER 1001\n" +
+                "\n"));
+    }
+
+    private static Build connectBuildFromGavAndUrl(String g, String a, String v, String url) {
+        MavenArtifact mvn = new MavenArtifactBuilder()
+                .withGroup(g)
+                .withArtifact(a)
+                .withVersion(v)
+                .build();
+        if (url != null) {
+            mvn.setRepository(url);
+        }
+
+        return new BuildBuilder()
+                .withPlugins(new PluginBuilder()
+                        .withName("my-connector-plugin")
+                        .withArtifacts(mvn)
+                        .build())
+                .build();
     }
 }